+++ Tests/Editor/Stubs/AttributeTest/ServiceWithPrivateResolveMethod.cs
@@ -1,0 +1,19 @@
+using TheRealIronDuck.Ducktion.Logging;
+
+namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs.AttributeTest
+{
+    public class ServiceWithPrivateResolveMethod
+    {
+        public SimpleService Simple { get; private set; }
+        public AnotherService Another { get; private set; }
+
+        [Attributes.Resolve]
+        private void ThisIsAMethodForSure(SimpleService simple, AnotherService another, DucktionLogger logger)
+        {
+            logger.Log(LogLevel.Debug, "I was called!");
+
+            Simple = simple;
+            Another = another;
+        }
+    }
+}